+++ packages/terra-clinical-result/src/result-time-header-cell/ResultTimeHeaderCell.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
-import VisuallyHiddenText from 'terra-visually-hidden-text';
 import styles from './ResultTimeHeaderCell.module.scss';
 
 const cx = classNamesBind.bind(styles);
@@ -62,8 +61,7 @@ const ResultTimeHeaderCell = (props) => {
       {...customProps}
       className={timeHeaderCellClassnames}
     >
-      <time className={dateClassnames}>{date}</time>
-      <VisuallyHiddenText text={'\u00A0'} />
+      <div className={dateClassnames}>{date}</div>
       <time className={cx('time')}>{time}</time>
     </th>
   );